kristian k: made login, logout, updated User DB(cash, user_type), updated base.html(with different user functions)
I have created the login, so a user is hardcoded, so when you click the login button you can login a user, if you want to change, you can change user to buyer. This needs a update in the future, where each button gives a different user. I have created logout, it is straight forward, you logout when you clik the button. I have updated the DB to store two different kinds of users, and each user have some cash when they start their account, we are going to use this to simulate buying items. I have updated the __init__ file, to login and logout a user, we use a loginmanager.